+++ trikot-viewmodels-declarative-flow/compose-flow/src/main/kotlin/com/mirego/trikot/viewmodels/declarative/compose/extensions/TextStyleResourceExtensions.kt
@@ -1,5 +1,6 @@
 package com.mirego.trikot.viewmodels.declarative.compose.extensions
 
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextStyle
@@ -10,14 +11,17 @@ import com.mirego.trikot.viewmodels.declarative.properties.VMDRichTextSpan
 import com.mirego.trikot.viewmodels.declarative.properties.VMDSpanStyleResourceTransform
 import com.mirego.trikot.viewmodels.declarative.properties.VMDTextStyleResource
 
+@Composable
 fun VMDTextStyleResource.textStyle(): TextStyle? =
     TrikotViewModelDeclarative.textStyleProvider.textStyleForResource(this)
 
+@Composable
 fun VMDTextViewModel.toAnnotatedString() = buildAnnotatedString {
     append(text)
     spans.forEach { addSpanStyle(it) }
 }
 
+@Composable
 private fun AnnotatedString.Builder.addSpanStyle(spanStyle: VMDRichTextSpan) {
     spanStyle.toComposeSpanStyle()?.let {
         addStyle(
@@ -28,6 +32,7 @@ private fun AnnotatedString.Builder.addSpanStyle(spanStyle: VMDRichTextSpan) {
     }
 }
 
+@Composable
 private fun VMDRichTextSpan.toComposeSpanStyle(): SpanStyle? =
     when (val currentTransform = transform) {
         is VMDSpanStyleResourceTransform -> currentTransform.textStyleResource.textStyle()?.toSpanStyle()