+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [VMD] Change parameters order for some Compose components
 - [VMD] Change VMDTextField's signature on Android
+- [VMD] Change VMDTextStyleProvider::textStyleForResource signature (now Composable)
 
 ### Updates